Fixed major bug: customers being counted doubly.
Sometimes item input can come in scattered (where banana appears more
than once in the customer's item list). For this reason, the customer
counts were initially incorrect, as it would increment customer counts
for an item each time the input was given.

Now, a list of customer names is kept for customer counts. The list
only expands if the customer's name is new, and the list's size
would be how many customers bought that item.